Remove Lazy for Scala 2.13
