+++ .github/workflows/ci.yml
@@ -29,7 +29,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        scala: [2.13.10, 2.12.16]
+        scala: [2.13.10, 2.12.17]
         java: [temurin@8]
         project: [root-spark31, root-spark32, root-spark33]
         exclude:
@@ -104,7 +104,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        scala: [2.12.16]
+        scala: [2.12.17]
         java: [temurin@8]
     runs-on: ${{ matrix.os }}
     steps:
@@ -160,7 +160,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        scala: [2.12.16]
+        scala: [2.12.17]
         java: [temurin@8]
     runs-on: ${{ matrix.os }}
     steps:
